@@ -11,44 +11,45 @@
   },
   "private": true,
   "dependencies": {
-    "@angular/animations": "~7.1.0",
-    "@angular/common": "~7.1.0",
-    "@angular/compiler": "~7.1.0",
-    "@angular/core": "~7.1.0",
-    "@angular/forms": "~7.1.0",
-    "@angular/platform-browser": "~7.1.0",
-    "@angular/platform-browser-dynamic": "~7.1.0",
-    "@angular/router": "~7.1.0",
-    "@ngrx/store": "^8.0.1",
-    "core-js": "^2.5.4",
-    "mobx": "^5.10.1",
-    "mobx-angular": "^3.0.1",
-    "ng-click-outside": "^5.1.1",
-    "ngx-pagination": "^4.0.0",
-    "rxjs": "~6.3.3",
-    "rxjs-compat": "^6.5.2",
-    "tslib": "^1.9.0",
-    "zone.js": "~0.8.26"
+    "@angular/animations": "^15.1.1",
+    "@angular/common": "^15.1.1",
+    "@angular/compiler": "^15.1.1",
+    "@angular/core": "^15.1.1",
+    "@angular/forms": "^15.1.1",
+    "@angular/platform-browser": "^15.1.1",
+    "@angular/platform-browser-dynamic": "^15.1.1",
+    "@angular/router": "^15.1.1",
+    "@ngrx/store": "^15.1.0",
+    "core-js": "^3.27.2",
+    "mobx": "^6.7.0",
+    "mobx-angular": "^4.6.0",
+    "ng-click-outside": "^9.0.1",
+    "ngx-pagination": "^6.0.3",
+    "rxjs": "^7.8.0",
+    "rxjs-compat": "^6.6.7",
+    "tslib": "^2.4.1",
+    "zone.js": "^0.12.0"
   },
   "devDependencies": {
-    "@angular-devkit/build-angular": "^0.12.1",
-    "@angular/cli": "~7.1.3",
-    "@angular/compiler-cli": "^8.0.0",
-    "@angular/language-service": "~7.1.0",
-    "@types/jasmine": "~2.8.8",
-    "@types/jasminewd2": "~2.0.3",
-    "@types/node": "~8.9.4",
-    "codelyzer": "~4.5.0",
-    "jasmine-core": "~2.99.1",
-    "jasmine-spec-reporter": "~4.2.1",
-    "karma": "^4.1.0",
-    "karma-chrome-launcher": "~2.2.0",
-    "karma-coverage-istanbul-reporter": "~2.0.1",
-    "karma-jasmine": "~1.1.2",
-    "karma-jasmine-html-reporter": "^0.2.2",
-    "protractor": "~5.4.0",
-    "ts-node": "~7.0.0",
-    "tslint": "~5.11.0",
-    "typescript": "~3.4.0"
+    "@angular-devkit/build-angular": "^15.1.2",
+    "@angular/cli": "^15.1.2",
+    "@angular/compiler-cli": "^15.1.1",
+    "@angular/language-service": "^15.1.1",
+    "@types/jasmine": "^4.3.1",
+    "@types/jasminewd2": "^2.0.10",
+    "@types/node": "^18.11.18",
+    "codelyzer": "^0.0.28",
+    "jasmine-core": "^4.5.0",
+    "jasmine-spec-reporter": "^7.0.0",
+    "karma": "^6.4.1",
+    "karma-chrome-launcher": "^3.1.1",
+    "karma-coverage-istanbul-reporter": "^3.0.3",
+    "karma-jasmine": "^5.1.0",
+    "karma-jasmine-html-reporter": "^2.0.0",
+    "protractor": "^7.0.0",
+    "ts-lint": "^4.5.1",
+    "ts-node": "^10.9.1",
+    "tslint": "^5.11.0",
+    "typescript": "^4.9.4"
   }
 }
